refactor(decorators): move _wrap_task into mixin
